add node at the end of a list.

+++ 0x12-singly_linked_lists/3-add_node_end.c
@@ -7,9 +7,9 @@
   *Return: length of string.
   */
 
-int _strlen(const char *str)
+size_t _strlen(const char *str)
 {
-	int i;
+	unsigned int i;
 
 	i = 0;
 	while (str[i] != '\0')
